+++ warning.go
@@ -2,22 +2,13 @@ package csv
 
 import "github.com/jszwec/csvutil"
 
-// Warning collect relevants informations about the process on the csv file
-type Warning map[string][]string
+// Warning collect relevants informations about the process on the csv file.
+type Warning map[string]string
 
-// NewWarning return a new warning
-// For read a warning it's possible to iterate over it like a map
+// NewWarning create a warning instance.
+// For read a warning it's possible to iterate over it like a map.
 func NewWarning() Warning {
-	return make(map[string][]string)
-}
-
-// Wrap relevants informations from the warn to the warning
-func (w *Warning) Wrap(warn Warning) {
-	for key, values := range warn {
-		for _, value := range values {
-			w.addValues(key, value)
-		}
-	}
+	return make(map[string]string)
 }
 
 func (w *Warning) unusedFields(decoder *csvutil.Decoder) {
@@ -26,15 +17,7 @@ func (w *Warning) unusedFields(decoder *csvutil.Decoder) {
 		if header[i] == "" {
 			continue
 		}
-		w.addValues(header[i], decoder.Record()[i])
-	}
-}
 
-func (w *Warning) addValues(key, value string) {
-	if _, ok := (*w)[key]; !ok {
-		(*w)[key] = []string{value}
-
-		return
+		(*w)[header[i]] = decoder.Record()[i]
 	}
-	(*w)[key] = append((*w)[key], value)
 }