+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # CHANGELOG
 
-## v0.0.2
+## v1.0.0
 
-- community standard addons
+- update codesytem
+- update linter
+- update ConfigDecoder interface to return an error and take a Decoder
+- fix tests
+- update example
+- update the decoder context to manipulate any instead of a string